project_root/requirements.txt
flask==2.0.1
pymysql==1.0.2
werkzeug==2.0.1
requests==2.26.0
beautifulsoup4==4.10.0
bcrypt==3.2.0
google-cloud-vision==2.4.1
Pillow==8.3.2
python-dotenv==0.19.0